SuperShop-v02/requirements/_base.txt
amqp==2.6.1
asgiref==3.2.10
attrs==20.3.0
Babel==2.8.0
backcall==0.2.0
billiard==3.6.3.0
braintree==3.59.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cssselect2==0.3.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-extensions==3.1.1
django-localflavor==3.0.1
django-parler==2.0.1
django-rosetta==0.9.3
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==6.0.0
flower==0.9.3
freezegun==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonschema==3.2.0
kombu==4.6.11
loguru==0.5.3
mypy-extensions==0.4.3
notifiers==1.2.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
polib==1.1.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycairo==1.20.0
pycparser==2.20
Pygments==2.7.4
Pyphen==0.10.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-stdnum==1.15
pytz==2020.1
redis==3.4.1
requests==2.24.0
rfc3987==1.3.8
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
tinycss2==1.0.2
tornado==5.1.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
